+++ CHANGELOG.md
@@ -26,7 +26,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - VAAPI may use hardware-accelerated tone mapping (when opencl accel is also available)
   - NVIDIA may use hardware-accelerated tone mapping (when vulkan accel and libplacebo filter are also available)
   - QSV may use hardware-accelerated tone mapping (when hardware decoding is used)
-  - In all other cases, HDR content will use a software pipeline and the linear algorithm
+  - In all other cases, HDR content will use a software pipeline
+  - The tonemap algorithm can be configured in the ffmpeg profile
 - Use hardware-accelerated padding with VAAPI
 - Add environment variable `ETV_DISABLE_VULKAN`
   - Any non-empty value will disable use of Vulkan acceleration and force software tonemapping